requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
tensorflow
matplotlib
wordcloud
joblib
comet-ml
fastapi
uvicorn
dvc
dvc-gs
